[BH-1486] Stop meditation time setting from increasing itself
Stop meditation time setting from increasing itself

+++ products/BellHybrid/apps/application-bell-meditation-timer/presenter/MeditationTimerPresenter.cpp
@@ -42,7 +42,7 @@ namespace app::meditation
     std::uint8_t MeditationTimerPresenter::getCurrentValue()
     {
         const auto value = settings->getValue(meditationDBRecordName, settings::SettingsScope::AppLocal);
-        auto defTimer    = utils::getNumericValue<std::uint8_t>(value);
+        auto defTimer    = utils::getNumericValue<std::uint32_t>(value);
         if (defTimer == emptyValue) {
             defTimer = defaultValue;
         }
